% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ghpages.R
\name{add_ghpages}
\alias{add_ghpages}
\title{Add GitHub Pages metadata}
\usage{
add_ghpages(meta = "echoverse", add_vignettes = TRUE, verbose = TRUE)
}
\arguments{
\item{meta}{A metadata \link[data.table]{data.table} produced
by \link[echodeps]{package_metadata}.
Alternatively, can be a character vector of packages to gather metadata for,
passed to \link[echodeps]{package_metadata}.}

\item{add_vignettes}{Search for any vignettes on the GitHub Pages websites,
and add them to a new column named "ghpages_vignettes".}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} of metadata.
}
\description{
For each package, list the links to the
dedicated GitHub Pages website(s).
If there are vignettes, add those as well.
If the GitHub Pages website for a package does not exist,
 \code{NULL} will be returned for that package.
}
\examples{
\dontrun{
meta <- add_ghpages()
}
}
