% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_colors.R
\name{construct_colors}
\alias{construct_colors}
\title{Construct colors}
\usage{
construct_colors(
  node_background = "#25355c",
  node_border = "#41c6c8",
  node_highlight = "#56ffff",
  node_hover_background = "blue",
  node_hover_border = node_border,
  node_font = "white",
  node_font_stroke = "rgba(103,115,141,.5)",
  node_shadow = "#537bcb",
  edge_color = node_highlight,
  edge_highlight = "#686ea6",
  edge_shadow = edge_highlight,
  tooltip_font = "#fff",
  tooltip_background = "rgba(0,0,0,.5)",
  tooltip_box_shadow = "rgba(247,247,247,0.5)",
  background = "transparent",
  save_background = background
)
}
\arguments{
\item{node_background}{node_background}

\item{node_border}{node_border}

\item{node_highlight}{node_highlight}

\item{node_hover_background}{node_hover_background}

\item{node_hover_border}{node_hover_border}

\item{node_font}{node_font}

\item{node_font_stroke}{node_font_stroke}

\item{node_shadow}{node_shadow}

\item{edge_color}{edge_color}

\item{edge_highlight}{edge_highlight}

\item{edge_shadow}{edge_shadow}

\item{tooltip_font}{tooltip_font}

\item{tooltip_background}{interaction_background}

\item{tooltip_box_shadow}{interaction_box_shadow}

\item{background}{background}

\item{save_background}{save_background}
}
\value{
Named list.
}
\description{
Construct a color mapping for \link[echodeps]{dep_graph_plot}.
}
\examples{
colors <- construct_colors()
}
