% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_graph_create.R
\name{dep_graph_create}
\alias{dep_graph_create}
\title{Create a dependency graph}
\usage{
dep_graph_create(
  pkg_name,
  deps = NULL,
  method = c("pkgnet", "github"),
  node_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pkg_name}{Package to search dependencies for.}

\item{deps}{A subset of of the main package's (\code{pkg_name} )
dependencies to include in the plot visualization.}

\item{method}{Method to create the dependency graph with:
\itemize{
\item{"pkgnet"}{Extracts all the R packages that the target R package depends
on using \link[pkgnet]{CreatePackageReport}}
\item{"github"}{Extracts all the GitHub repositories that depend on the
respective GitHub repository of the target R package
using \link[echogithub]{github_dependents}}
}
If more than one option is provided, only the first is used.}

\item{node_size}{Strategy for setting node sizes. Can be one of:
\itemize{
\item{NULL : }{Sets the target package node's size to 40 and the rest to 30}.
\item{<numeric> : }{User-provided node size.
Sets all nodes to the same size.}
\item{<character> : }{User-provided metadata column name.
Scales node size to a vector of numeric values in the node metadata.}
}}

\item{verbose}{Print messages.}
}
\description{
Create a dependency graph.
}
\examples{
dgc_out <- dep_graph_create(pkg_name = "rworkflows",
                            method = "github")
}
