% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graphs.R
\name{example_graphs}
\alias{example_graphs}
\title{Example graphs}
\usage{
example_graphs(
  n = c(4, 5, 10),
  p = 0.5,
  colors = pals::viridis(length(n)),
  seed = 2023
)
}
\arguments{
\item{n}{Number of nodes for each graph.}

\item{p}{The probability of nodes connecting between graphs (on a 0-1 scale).}

\item{colors}{Color for each graph.}

\item{seed}{Random seed.}
}
\value{
A list of  \link[tidygraph]{tbl_graph} objects.
}
\description{
Generate a random list of  \link[tidygraph]{tbl_graph} objects.
}
\examples{
grs <- example_graphs()
}
