% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_metadata.R
\name{package_metadata}
\alias{package_metadata}
\title{Gather package metadata}
\usage{
package_metadata(
  pkgs = "echoverse",
  fields = c("Package", "Title", "Description", "URL", "Version", "Depends", "Imports",
    "Suggests", "Remotes", "SystemRequirements"),
  verbose = TRUE
)
}
\arguments{
\item{pkgs}{Packages to gather metadata for.}

\item{fields}{DESCRIPTION file fields.
If a field is missing for a given package, will fill with \code{NULL}.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} of metadata.
}
\description{
Gather metadata from the DESCRIPTION file of each R package and construct a
\link[data.table]{data.table}.
 By default, returns metadata for all
 \href{https://github.com/topics/echoverse}{\code{echoverse}} packages.
}
\examples{
meta <- echodeps::package_metadata()
}
