% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_graph_join.R
\name{recursive_graph_join}
\alias{recursive_graph_join}
\title{Recursively join a list of \link[tidygraph]{tbl_graph}s}
\usage{
recursive_graph_join(grs, by = "name")
}
\arguments{
\item{grs}{a list of \link[tidygraph]{tbl_graph} objects.}

\item{by}{A join specification created with \code{\link[dplyr:join_by]{join_by()}}, or a character
vector of variables to join by.

If \code{NULL}, the default, \verb{*_join()} will perform a natural join, using all
variables in common across \code{x} and \code{y}. A message lists the variables so
that you can check they're correct; suppress the message by supplying \code{by}
explicitly.

To join on different variables between \code{x} and \code{y}, use a \code{\link[dplyr:join_by]{join_by()}}
specification. For example, \code{join_by(a == b)} will match \code{x$a} to \code{y$b}.

To join by multiple variables, use a \code{\link[dplyr:join_by]{join_by()}} specification with
multiple expressions. For example, \code{join_by(a == b, c == d)} will match
\code{x$a} to \code{y$b} and \code{x$c} to \code{y$d}. If the column names are the same between
\code{x} and \code{y}, you can shorten this by listing only the variable names, like
\code{join_by(a, c)}.

\code{\link[dplyr:join_by]{join_by()}} can also be used to perform inequality, rolling, and overlap
joins. See the documentation at \link[dplyr:join_by]{?join_by} for details on
these types of joins.

For simple equality joins, you can alternatively specify a character vector
of variable names to join by. For example, \code{by = c("a", "b")} joins \code{x$a}
to \code{y$a} and \code{x$b} to \code{y$b}. If variable names differ between \code{x} and \code{y},
use a named character vector like \code{by = c("x_a" = "y_a", "x_b" = "y_b")}.

To perform a cross-join, generating all combinations of \code{x} and \code{y}, see
\code{\link[dplyr:cross_join]{cross_join()}}.}
}
\value{
a single tidygraph object
}
\description{
Joins a list of graphs into a single graph.
This recursively implements \link[tidygraph]{graph_join}.
}
\examples{
set.seed(2023)
grs <- example_graphs()
g <- recursive_graph_join(grs)
plot(g, edge.label=igraph::E(g)$color)
}
