% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_graph_create.R
\name{revdep_graph_create}
\alias{revdep_graph_create}
\title{Create a reverse dependency graph}
\source{
\href{https://stackoverflow.com/q/38428945}{igraph::union issue}
}
\usage{
revdep_graph_create(
  pkg,
  exclude = NULL,
  method_seed = c("github", "devtools"),
  method = c("devtools", "github"),
  recursive = FALSE,
  node_size = NULL,
  add_metadata = TRUE,
  degrees = 2,
  verbose = TRUE
)
}
\arguments{
\item{pkg}{Package to search dependencies for.}

\item{exclude}{A subset of of the main package's (\code{pkg} )
dependencies to exclude in the graph.}

\item{method_seed}{Method to create the initial dependency graph with:
\itemize{
\item{"pkgnet" (\code{reverse=FALSE} only)}{
Extracts all the R packages that the target R package depends
on using \link[pkgnet]{CreatePackageReport}}
\item{"github" (\code{reverse=FALSE or TRUE})}{
Extracts all the GitHub repositories that depend on the
respective GitHub repository of the target R package
using \link[echogithub]{github_dependents}}
\item{"devtools" (\code{reverse=TRUE}) only}{
Extracts reverse dependencies using \link[devtools]{revdep}.}
}
If more than one option is provided, only the first is used.}

\item{method}{Seed method for extracting first-order reverse dependencies
of \code{pkg}.}

\item{recursive}{If \code{TRUE} look for full set of recursive dependencies.}

\item{node_size}{Strategy for setting node sizes. Can be one of:
\itemize{
\item{NULL : }{Sets the target package node's size to 40 and the rest to 30}.
\item{<numeric> : }{User-provided node size.
Sets all nodes to the same size.}
\item{<character> : }{User-provided metadata column name.
Scales node size to a vector of numeric values in the node metadata.}
}}

\item{add_metadata}{Add metadata to the graph
using \link[echogithub]{github_metadata}.}

\item{degrees}{The number of degrees out from the main \code{pkg} node to
extend the dependency graph.}

\item{verbose}{Print messages.}
}
\value{
Named list.
}
\description{
Create a reverse dependency graph.
}
\examples{
dgc_out <- revdep_graph_create(pkg = "rworkflows")
}
