% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_graph_create_devtools.R
\name{revdep_graph_create_devtools}
\alias{revdep_graph_create_devtools}
\title{Create a reverse dependency graph: devtools}
\usage{
revdep_graph_create_devtools(
  pkgs,
  pkg_target = NULL,
  exclude = NULL,
  recursive = FALSE,
  add_metadata = TRUE,
  use_basename = TRUE,
  sep = "/\\n",
  node_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pkgs}{Package names.}

\item{exclude}{A subset of of the main package's (\code{pkg} )
dependencies to exclude in the graph.}

\item{recursive}{If \code{TRUE} look for full set of recursive dependencies.}

\item{add_metadata}{Add metadata to the graph
using \link[echogithub]{github_metadata}.}

\item{use_basename}{Only use the repo name to label each node.}

\item{sep}{Character separator between owner and repo, to label each node.}

\item{node_size}{Strategy for setting node sizes. Can be one of:
\itemize{
\item{NULL : }{Sets the target package node's size to 40 and the rest to 30}.
\item{<numeric> : }{User-provided node size.
Sets all nodes to the same size.}
\item{<character> : }{User-provided metadata column name.
Scales node size to a vector of numeric values in the node metadata.}
}}

\item{verbose}{Print messages.}
}
\value{
igraph
}
\description{
Create a reverse dependency graph.
}
\keyword{internal}
