% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_graph.R
\name{subset_graph}
\alias{subset_graph}
\title{Subset graph}
\usage{
subset_graph(g, include = NULL, exclude = NULL, verbose = TRUE)
}
\arguments{
\item{g}{A graph of class \link[igraph]{igraph}.
\emph{Warning:} Large graphs may be slow to interact with.}

\item{include}{Names of nodes to include.}

\item{exclude}{Names of nodes to exclude.}

\item{verbose}{Print messages.}
}
\value{
Subgraph.
}
\description{
Subset a graph by nodes to include or exclude.
Automatically omits \code{include}/\code{exclude} nodes
 that don't exist in the graph to avoid errors.
}
\examples{
dgc_out <- dep_graph_create(pkg_name = "rworkflows",
                            method = "github",
                            node_size = "clones_uniques")
g <- dgc_out$graph
exclude <- grep("actions-marketplace-validations",
                names(igraph::V(g)),value = TRUE)
g2 <- subset_graph(g = g, exclude = exclude)
}
