% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visnet_save.R
\name{visnet_save}
\alias{visnet_save}
\title{visNetwork save}
\usage{
visnet_save(
  url,
  file = gsub("\\\\.html", ".png", url),
  res = 8,
  delay = 0.2,
  zoom = 1,
  vwidth = 992 * res,
  vheight = 744 * res,
  geometry = NULL,
  run_shrink = FALSE
)
}
\arguments{
\item{url}{A vector of URLs to visit. If multiple URLs are provided, it will
load and take screenshots of those web pages in parallel.}

\item{file}{A vector of names of output files. Should end with an image file
type (\code{.png}, \code{.jpg}, \code{.jpeg}, or \code{.webp}) or
\code{.pdf}. If several screenshots have to be taken and only one filename
is provided, then the function appends the index number of the screenshot
to the file name. For PDF output, it is just like printing the page to PDF
in a browser; \code{selector}, \code{cliprect}, \code{expand}, and
\code{zoom} will not be used for PDFs.}

\item{res}{Resolution multiplier.}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a
longer delay is needed for all assets to display properly.}

\item{zoom}{A number specifying the zoom factor. A zoom factor of 2 will
result in twice as many pixels vertically and horizontally. Note that using
2 is not exactly the same as taking a screenshot on a HiDPI (Retina)
device: it is like increasing the zoom to 200% in a desktop browser and
doubling the height and width of the browser window. This differs from
using a HiDPI device because some web pages load different,
higher-resolution images when they know they will be displayed on a HiDPI
device (but using zoom will not report that there is a HiDPI device).}

\item{vwidth}{Viewport width. This is the width of the browser "window".}

\item{vheight}{Viewport height This is the height of the browser "window".}

\item{geometry}{Scaling specification. Can be a percent, as in \code{"50\%"},
or pixel dimensions like \code{"120x120"}, \code{"120x"}, or \code{"x120"}.
Any valid ImageMagick geometry specification can be used. If \code{filename}
contains multiple images, this can be a vector to specify distinct sizes
for each image.}

\item{run_shrink}{Run \link[webshot2]{shrink} on the png.}
}
\value{
png file path
}
\description{
Save a visNetwork object as a png.
}
\examples{
save_path <- tempfile(fileext = ".html")
res <- dep_graph(pkg = "echoverse",
                 layout=layout_star,
                 save_path=save_path)
visnet_save(url = "https://datastorm-open.github.io/visNetwork/")
}
