% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_graph_create_github.R
\name{dep_graph_create_github}
\alias{dep_graph_create_github}
\title{Create a dependency graph: GitHub}
\usage{
dep_graph_create_github(
  pkg,
  exclude = NULL,
  add_traffic = TRUE,
  sep = "/\\n",
  use_basename = TRUE,
  node_size = NULL,
  reverse = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pkg}{Package to search dependencies for.}

\item{exclude}{A subset of of the main package's (\code{pkg} )
dependencies to exclude in the graph.}

\item{add_traffic}{Add traffic metadata 
with \link[echogithub]{github_traffic}.}

\item{sep}{Character separator between owner and repo, to label each node.}

\item{use_basename}{Only use the repo name to label each node.}

\item{node_size}{Strategy for setting node sizes. Can be one of:
\itemize{
\item{NULL : }{Sets the target package node's size to 40 and the rest to 30}.
\item{<numeric> : }{User-provided node size.
Sets all nodes to the same size.}
\item{<character> : }{User-provided metadata column name.
Scales node size to a vector of numeric values in the node metadata.}
}}

\item{reverse}{\itemize{
\item{\code{FALSE} (default)}{Create a forward dependency graph
(packages that \code{pkg} depends on,
and the packages that they depend on).}
\item{\code{TRUE}}{Create a reverse dependency graph
 (packages that depend on \code{pkg},
 and the packages that depend on them).}
}}

\item{verbose}{Print messages.}
}
\description{
Create a dependency graph: GitHub.
}
