% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_graph_plot.R
\name{dep_graph_plot}
\alias{dep_graph_plot}
\title{Dependency graph plot}
\usage{
dep_graph_plot(
  g,
  pkg,
  shape = c("image", "hexagon"),
  image = "hex_sticker",
  layout = layout_igraph,
  hide_tooltip = TRUE,
  show_plot = list(r = TRUE, browser = TRUE),
  save_path = file.path(tempdir(), paste0(basename(pkg), ".dep_graph.html")),
  width = NULL,
  height = NULL,
  colors = construct_colors(),
  font_face = "Tahoma",
  use_basename = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{g}{A graph of class \link[igraph]{igraph}.
\emph{Warning:} Large graphs may be slow to interact with.}

\item{pkg}{Package to search dependencies for.}

\item{shape}{: String. Default to 'ellipse'. The shape defines what the node looks like. There are two types of nodes. One type has the label inside of it and the other type has the label underneath it. The types with the label inside of it are: ellipse, circle, database, box, text. The ones with the label outside of it are: image, circularImage, diamond, dot, star, triangle, triangleDown, hexagon, square and icon.}

\item{image}{: List or String. Default to undefined. When the shape is set to image or circularImage, this option should be the URL to an image. If the image cannot be found, the brokenImage option can be used.
\itemize{
  \item{"unselected"}{ : String. Unselected (default) image URL.}
  \item{"selected"}{ : String. Selected image URL.}
}}

\item{layout}{\pkg{visNetwork} layout function
(e.g. \link[visNetwork]{visLayout} or \link[visNetwork]{visIgraphLayout}
to specify plot layout. The function must take two arguments:
"graph" and "pkg".
See \code{echodeps::layout_star} for an example.
Alternatively, you may set \code{layout=NULL} to use the
default force-directed \link[visNetwork]{visIgraph} layout, "layout_nicely".}

\item{hide_tooltip}{Hide the tooltips by default, and only show them when
the cursor is hovering over the respective node.}

\item{show_plot}{A named list with two items:
\itemize{
\item{r}{Whether to show the dependency graph in R/Rstudio
(will \emph{not} show user-specific \code{background} color)}.
\item{browser}{Whether to show the dependency graph in default web browser
(will show user-specific \code{background} color)}.
}}

\item{save_path}{Path to save the plot to, as an interactive,
self-container HTML file.}

\item{width}{: String. Default to "100\%". The width of the network in pixels or as a percentage.}

\item{height}{: String. Default to "100\%". The height of the network in pixels or as a percentage.}

\item{colors}{Colors for each of the graph plot elements.
Can use the \link[echodeps]{construct_colors} function to help create this.}

\item{font_face}{Font to use for plot text.}

\item{use_basename}{Only use the repo name to label each node.}

\item{verbose}{Print messages.}
}
\description{
Plot a dependency graph between a set of R packages
as an interactive network.
}
\examples{
dgc_out <- dep_graph_create(pkg = "rworkflows",
                            method = "github")
vis <- dep_graph_plot(g = dgc_out$subgraph,
                      pkg = dgc_out$pkg)
}
