% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_package.R
\name{is_r_package}
\alias{is_r_package}
\title{Is R package}
\usage{
is_r_package(pkg, owner = NULL, verbose = TRUE)
}
\arguments{
\item{pkg}{Package name.}

\item{owner}{Name of the owner of the R packages'
GitHub repository (optional).}

\item{verbose}{Print messages.}
}
\value{
boolean
}
\description{
Check whether a GitHub repository is for an R package or not.
}
\examples{
is_pkg <- is_r_package(pkg="stats")
}
