% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_igraph.R
\name{layout_igraph}
\alias{layout_igraph}
\title{\pkg{igraph} layout}
\usage{
layout_igraph(graph, layout = "layout_nicely", ...)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{layout}{: Character Name of igraph layout function to use. Default to "layout_nicely"}

\item{...}{Additional arguments passed to
\link[visNetwork]{visIgraphLayout}.}
}
\description{
Layout function uses any \link[visNetwork]{visIgraphLayout}
}
