% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_star.R
\name{layout_star}
\alias{layout_star}
\title{Star layout}
\usage{
layout_star(graph, pkg, ...)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{pkg}{Name of the node to place at the center of the plot
(e.g. "echolocatoR").}

\item{...}{Additional arguments passed to
\link[visNetwork]{visIgraphLayout}.}
}
\description{
Default layout function that arranges nodes in a star pattern with
one node (\code{pkg}) at its center.
}
