% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_graph_create.R
\name{revdep_graph_create}
\alias{revdep_graph_create}
\title{Create a reverse dependency graph}
\source{
\href{https://stackoverflow.com/q/38428945}{igraph::union issue}
}
\usage{
revdep_graph_create(
  pkg,
  exclude = NULL,
  method_seed = c("github", "devtools"),
  method = c("devtools", "github"),
  use_basename = TRUE,
  sep = "/\\n",
  recursive = FALSE,
  node_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pkg}{Package to search dependencies for.}

\item{exclude}{A subset of of the main package's (\code{pkg} )
dependencies to exclude in the graph.}

\item{method_seed}{Method to get the initial layer of
reverse dependencies with.}

\item{method}{Method to create the dependency graph with:
\itemize{
\item{"pkgnet"}{Extracts all the R packages that the target R package depends
on using \link[pkgnet]{CreatePackageReport}}
\item{"github"}{Extracts all the GitHub repositories that depend on the
respective GitHub repository of the target R package
using \link[echogithub]{github_dependents}}
}
If more than one option is provided, only the first is used.}

\item{use_basename}{Only use the repo name to label each node.}

\item{sep}{Character separator between owner and repo, to label each node.}

\item{recursive}{If \code{TRUE} look for full set of recursive dependencies.}

\item{node_size}{Strategy for setting node sizes. Can be one of:
\itemize{
\item{NULL : }{Sets the target package node's size to 40 and the rest to 30}.
\item{<numeric> : }{User-provided node size.
Sets all nodes to the same size.}
\item{<character> : }{User-provided metadata column name.
Scales node size to a vector of numeric values in the node metadata.}
}}

\item{verbose}{Print messages.}
}
\description{
Create a reverse dependency graph.
}
\examples{
dgc_out <- revdep_graph_create(pkg = "rworkflows")
}
